code review + renaming
