+++ src/components/Invoices/InvoiceOverview/InvoiceOverview.tsx
@@ -1,12 +1,12 @@
 import { VStack } from '../../ui/Stack/Stack'
-import { InvoicesTable } from '../InvoicesTable/InvoicesTable'
+import { InvoiceTable } from '../InvoiceTable/InvoiceTable'
 import { InvoiceSummaryStats } from '../InvoiceSummaryStats/InvoiceSummaryStats'
 
 export const InvoiceOverview = () => {
   return (
     <VStack>
       <InvoiceSummaryStats />
-      <InvoicesTable />
+      <InvoiceTable />
     </VStack>
   )
 }
