+++ src/components/ui/Table/table.scss
@@ -19,6 +19,7 @@
 }
 
 .Layer__UI__Table-Column {
+  box-sizing: border-box;
   padding: var(--spacing-xs) var(--spacing-sm);
   font-size: var(--text-sm);
   font-weight: var(--font-weight-normal);
