+++ src/providers/InvoiceStore/InvoiceStoreProvider.tsx
@@ -1,7 +1,7 @@
 import { useState, createContext, useContext, useMemo, type PropsWithChildren } from 'react'
 import { createStore, useStore } from 'zustand'
 import type { Invoice } from '../../features/invoices/invoiceSchemas'
-import { ALL_OPTION, type InvoiceStatusOption } from '../../components/Invoices/InvoicesTable/InvoicesTable'
+import { ALL_OPTION, type InvoiceStatusOption } from '../../components/Invoices/InvoiceTable/InvoiceTable'
 import type { InvoiceFormMode } from '../../components/Invoices/InvoiceForm/InvoiceForm'
 import { UpsertInvoiceMode } from '../../features/invoices/api/useUpsertInvoice'
