+++ src/styles/index.scss
@@ -31,7 +31,7 @@
 @forward '../components/LinkedAccounts/BasicLinkedAccount/basicLinkedAccount';
 @forward '../components/BaseDetailView/baseDetailView';
 @forward '../components/Bills/bills';
-@forward '../components/Invoices/InvoicesTable/invoicesTable';
+@forward '../components/Invoices/InvoiceTable/invoiceTable';
 @forward '../components/Invoices/InvoiceDetail/invoiceDetail';
 @forward '../components/Invoices/InvoiceForm/invoiceForm';
 @forward '../components/Invoices/InvoiceSummaryStats/invoiceSummaryStats';
