+++ src/utils/bigDecimalUtils.ts
@@ -34,7 +34,7 @@ export const convertBigDecimalToCents = (amount: BD.BigDecimal): number => {
   return Number(rounded.value)
 }
 
-export const convertBigIntCentsToBigDecimal = (cents: bigint) => {
+export const convertBigIntCentsToBigDecimal = (cents: bigint): BD.BigDecimal => {
   // Create a BigDecimal representing the cents
   const decimalCents = BD.fromBigInt(cents)
